rename(apps/app): Tab 폴더 내 스크린 파일 네이밍 변경
