+++ apps/app/src/navigators/TabNavigator.tsx
@@ -1,7 +1,7 @@
 import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
-import CreateDiaryScreen from '../screens/Tab/CreateDiary';
-import HomeScreen from '../screens/Tab/Home';
-import SettingScreen from '../screens/Tab/Setting';
+import CreateDiaryScreen from '../screens/Tab/CreateDiaryScreen';
+import HomeScreen from '../screens/Tab/HomeScreen';
+import SettingScreen from '../screens/Tab/SettingScreen';
 import {useSafeAreaInsets} from 'react-native-safe-area-context';
 import {theme} from '@greeenai/design-tokens';
 import {RouteProp} from '@react-navigation/native';
