+++ apps/app/src/screens/Tab/CreateDiaryScreen.tsx
